Refactors the $all stream for reading.
Code was consolidated from the readallforward and readallbackward command(s) to re-utilize code.
